aws_textract/requirements.txt
asgiref==3.5.2
backports.zoneinfo==0.2.1
boto3==1.24.12
botocore==1.27.12
certifi==2022.6.15
charset-normalizer==2.0.12
Django==4.0.5
django-storages==1.12.3
idna==3.3
jmespath==1.0.1
lxml==4.9.0
numpy==1.22.4
pandas==1.4.2
Pillow==9.1.1
psycopg2==2.9.3
pymongo==4.1.1
pyspark==3.2.1
python-dateutil==2.8.2
python-dotenv==0.20.0
pytz==2022.1
requests==2.28.0
s3transfer==0.6.0
six==1.16.0
sqlparse==0.4.2
textract-trp==0.1.3
urllib3==1.26.9
